Move exiting logic from main window module to a separate helper module.
